requirements.txt
pymongo>=4.6.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
click>=8.1.7

